Factor index into module
